Refactor game state helpers into dedicated hooks
