+++ packages/web/src/state/actionTypes.ts
@@ -1,0 +1,5 @@
+export interface Action {
+	id: string;
+	name: string;
+	system?: boolean;
+}
