+++ packages/web/src/state/phaseTypes.ts
@@ -1,0 +1,5 @@
+export type PhaseStep = {
+	title: string;
+	items: { text: string; italic?: boolean; done?: boolean }[];
+	active: boolean;
+};
